Apollo: Release source code for 55.4 (1504)

+++ libwallet/service/model/security_card_metadata.go
@@ -4,7 +4,7 @@ type SecurityCardMetadataJson struct {
 	GlobalPublicKeyInHex string `json:"globalPublicKeyInHex"`
 	CardVendorInHex      string `json:"cardVendorInHex"`
 	CardModelInHex       string `json:"cardModelInHex"`
-	FirmwareVersion      int    `json:"firmwareVersion"`
-	UsageCount           int    `json:"usageCount"`
+	FirmwareVersion      uint16 `json:"firmwareVersion"`
+	UsageCount           uint16 `json:"usageCount"`
 	LanguageCodeInHex    string `json:"languageCodeInHex"`
 }
